refactor: ShoppingAppContainer를 ContainerBuilder로 분리
