+++ app/src/main/java/woowacourse/shopping/di/DiBindings.kt
@@ -5,7 +5,7 @@ import woowacourse.shopping.data.ProductRepositoryImpl
 import woowacourse.shopping.domain.CartRepository
 import woowacourse.shopping.domain.ProductRepository
 
-fun installAllBindings(container: ShoppingAppContainer) {
-    container.register(ProductRepository::class) { ProductRepositoryImpl() }
-    container.register(CartRepository::class) { CartRepositoryImpl() }
+fun installAllBindings(builder: ContainerBuilder) {
+    builder.register(ProductRepository::class) { ProductRepositoryImpl() }
+    builder.register(CartRepository::class) { CartRepositoryImpl() }
 }
